// clients/javascript/lib/gen_types/CalendarEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CalendarEventReminder } from './CalendarEventReminder'
import type { CalendarEventStatus } from './CalendarEventStatus'
import type { ID } from './ID'
import type { RRuleOptions } from './RRuleOptions'

export type CalendarEvent = {
  id: ID
  parentId: string | null
  title: string | null
  description: string | null
  location: string | null
  allDay: boolean
  status: CalendarEventStatus
  startTime: Date
  duration: bigint
  busy: boolean
  endTime: Date
  created: bigint
  updated: bigint
  recurrence: RRuleOptions | null
  exdates: Array<Date>
  calendarId: ID
  userId: ID
  accountId: ID
  reminders: Array<CalendarEventReminder>
  serviceId: ID | null
  metadata: Record<string, string>
}